i parsing working/cd/pwd/

+++ minishell/free.c
@@ -6,12 +6,14 @@
 /*   By: juliencros <juliencros@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/10 13:56:36 by juliencros        #+#    #+#             */
-/*   Updated: 2023/08/21 16:59:36 by juliencros       ###   ########.fr       */
+/*   Updated: 2023/09/20 15:29:16 by juliencros       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "free.h"
 #include <unistd.h>
+#include <stdio.h>
+#include "str.h"
 
 void	ft_free_tab(char **tab);
 
@@ -24,13 +26,19 @@ void	ft_free_subcommands(t_subcommand *subcommand)
 			close(subcommand->in_fd);
 			unlink(".here_doc_fd");
 		}
-		if (subcommand->in_fd != 0 && subcommand->in_fd != -1 || subcommand->out_fd != 1)
+		if ((subcommand->in_fd != 0) && (subcommand->in_fd != -1) && (!subcommand->is_heredoc) && (subcommand->out_fd != subcommand->in_fd))
 			close(subcommand->in_fd);
 		if (subcommand->out_fd != 1)
 		{
 			close(subcommand->out_fd);
-			unlink("tmp");
+			if (ft_strncmp(subcommand->out_file_name, "tmp", 3) == 0)
+				unlink("tmp");
+			
 		}
+		if (subcommand->out_file_name != NULL)
+			free(subcommand->out_file_name);	
+		if (subcommand->path != NULL)
+			free(subcommand->path);
 		if (subcommand->args != NULL)
 			ft_free_tab(subcommand->args);
 		subcommand = subcommand->next;
